Server/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
PyJWT==2.1.0
Werkzeug==2.2.2